Replace std::vector<int> with IntegerVector
In the RCpp function pairwise_geno_id() I had some
some code that must have been written before
IntegerVector had a push_back() method. I was
using std::vector<int> for that but I was not
properly casting it back to IntegerVector for
return.  I hope this fixes some issues we are
having with this function.